+++ backends/simplego/exec_special_ops_float16.go
@@ -9,7 +9,6 @@ import (
 	"unsafe"
 
 	"github.com/gomlx/gomlx/pkg/core/dtypes"
-	"github.com/gomlx/gomlx/pkg/core/dtypes/bfloat16"
 	"github.com/gomlx/gomlx/pkg/support/xslices"
 	"github.com/x448/float16"
 )
@@ -89,96 +88,6 @@ func execReduceProductFloat16(operand, output *Buffer, it *reduceOutputIterator,
 }
 
 // Float16 conversion functions
-
-func execConvertDTypeFromFloat16[_ float16.Float16, ToT PODNumericConstraints](operand, output *Buffer) {
-	operandFlat := operand.flat.([]float16.Float16)
-	outputFlat := output.flat.([]ToT)
-	for idx, value := range operandFlat {
-		outputFlat[idx] = ToT(value.Float32())
-	}
-}
-
-func execConvertDTypeToFloat16[FromT PODNumericConstraints, _ float16.Float16](operand, output *Buffer) {
-	operandFlat := operand.flat.([]FromT)
-	outputFlat := output.flat.([]float16.Float16)
-	for idx, value := range operandFlat {
-		outputFlat[idx] = float16.Fromfloat32(float32(value))
-	}
-}
-
-func execConvertDTypeFloat16ToBool(operand, output *Buffer) {
-	operandFlat := operand.flat.([]float16.Float16)
-	outputFlat := output.flat.([]bool)
-	for idx, value := range operandFlat {
-		outputFlat[idx] = value.Float32() != 0
-	}
-}
-
-func execConvertDTypeBoolToFloat16(operand, output *Buffer) {
-	operandFlat := operand.flat.([]bool)
-	outputFlat := output.flat.([]float16.Float16)
-	zero, one := float16.Fromfloat32(0), float16.Fromfloat32(1)
-	for idx, value := range operandFlat {
-		if value {
-			outputFlat[idx] = one
-		} else {
-			outputFlat[idx] = zero
-		}
-	}
-}
-
-func execConvertDTypeFloat16ToBFloat16(operand, output *Buffer) {
-	operandFlat := operand.flat.([]float16.Float16)
-	outputFlat := output.flat.([]bfloat16.BFloat16)
-	for idx, value := range operandFlat {
-		outputFlat[idx] = bfloat16.FromFloat32(value.Float32())
-	}
-}
-
-func execConvertDTypeBFloat16ToFloat16(operand, output *Buffer) {
-	operandFlat := operand.flat.([]bfloat16.BFloat16)
-	outputFlat := output.flat.([]float16.Float16)
-	for idx, value := range operandFlat {
-		outputFlat[idx] = float16.Fromfloat32(value.Float32())
-	}
-}
-
-func init() {
-	// Register Float16 conversion functions
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Int8, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, int8])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Int16, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, int16])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Int32, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, int32])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Int64, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, int64])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Uint8, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, uint8])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Uint16, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, uint16])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Uint32, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, uint32])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Uint64, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, uint64])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Float32, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, float32])
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Float64, priorityTyped, execConvertDTypeFromFloat16[float16.Float16, float64])
-
-	convertDTypePairMap.Register(dtypes.Int8, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[int8, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Int16, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[int16, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Int32, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[int32, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Int64, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[int64, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Uint8, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[uint8, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Uint16, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[uint16, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Uint32, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[uint32, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Uint64, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[uint64, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Float32, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[float32, float16.Float16])
-	convertDTypePairMap.Register(dtypes.Float64, dtypes.Float16, priorityTyped, execConvertDTypeToFloat16[float64, float16.Float16])
-
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Bool, priorityTyped, execConvertDTypeFloat16ToBool)
-	convertDTypePairMap.Register(dtypes.Bool, dtypes.Float16, priorityTyped, execConvertDTypeBoolToFloat16)
-
-	// Float16 <-> BFloat16 conversion
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.BFloat16, priorityTyped, execConvertDTypeFloat16ToBFloat16)
-	convertDTypePairMap.Register(dtypes.BFloat16, dtypes.Float16, priorityTyped, execConvertDTypeBFloat16ToFloat16)
-
-	// Float16 <-> Float16 identity
-	convertDTypePairMap.Register(dtypes.Float16, dtypes.Float16, priorityTyped, func(operand, output *Buffer) {
-		copy(output.flat.([]float16.Float16), operand.flat.([]float16.Float16))
-	})
-}
 
 // Float16 buffer operations