+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@
   - Added dedup of computation nodes (aka. "common subexpression elimination" CSE) (thx @timkaye11)
     - ~6% speedup for CSI-Adult demo training. 
   - DotGeneral: Pre-blocking of the blocked path, which may lead to deduplication of blocking nodes.
+  - DotGeneral: Added smallMatMul execution path, optimized for small matrix multiplications (thx @timkaye11)
 - New package `bucketing`:
   - Tools to manage bucketing of tensors (or anything else) -- thx @ajroetker
 - Package `dtypes`:
@@ -18,6 +19,8 @@
 - Package `graph`:
   - Added `Unpack()` and `Pack()` for sub-byte dtypes.
   - Added `Function` concept (and support for closures) and the `Function.Call` operation.
+  - Control Flow: Added `While` and `If` operations.
+  - Order operations: Added `Sort`, `SortFunc`, `TopK`, `BottomK`.
   
 # v0.26.0: Using the new github.com/gomlx/go-xla library. Added linux/arm64 and windows/amd64 support for XLA CPU.